Oscilloscope: Move the measurement selection logic to MeasureSettings
class Measure owns the measurement for a channel.
class MeasureSettings handles the measurement selection and does not
change any data of the measurement.
class Oscilloscope uses the selection (managed by class MeasureSettings)
to display the selected measurements. The measurements that are on the list
are being modified (enabled if on the list, disabled otherwise) by the class
Oscilloscope.

Signed-off-by: Dan <Dan.Nechita@analog.com>