Oscilloscope: Display buffer size and sample rate near the plot canvas
This displays the number of samples and the sample rate at which the capture process is currently set and will not always apply to the data being displayed on the screen. For example if it takes some time for data to be received by the instrument and time base gets changed before data arrives then the label will display the latter settings and not the ones that apply to the data - which is not ideal and needs to be addressed. However this is useful in most situations where data latency is minimal and also to know the current sampling rate.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>