Oscilloscope: Restructure the trigger mechanism
Trigger positions can exceed the left/right limits of the visible plot.

When position exceeds the right limit, the buffer size increases
(visible samples + hidden samples on the right). Once the buffer size
exceeds the size of the trigger buffer (which is 8k) the samples rate is
reduced by one step so that we can avoid limiting the position of the
trigger that can be set to the right.

When position exceeds the left limit, the buffer size increases (visible
samples + hidden samples on the left). The limit of the buffer size is
currently set to 500k and cannot be exceeded thus limiting the position
that can be set to the left.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>