src/buffer_previewer: A new widget that provides a top view of a buffer
The widget composition:
 - a waveform that illustrates a buffer
 - a rectangle that highlights a part of the entire widget and a part of
the waveform (if the highlight area intersects any of its points)
 - a cursor

The widths and positions of the above elements can be set anywhere
between 0% and 100% of the entire width of the widget. The cursor does
not have a width.

The widget can be used in acquisition instruments to display information
about the buffer length, position, the visible/hidden part and the
trigger position.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>