src/buffer_previewer: Avoid highlighted color mixing with default color
Because antialiasing is enabled while drawing the waveforms, the colors of
the default waveform and of the highlighted waveform get mixed together.

Avoid this by first drawing the default waveform then the highlight
rectangle on top of it and finally draw the highlighted waveform on top
of the rectangle surface.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>