Add ApiObject class
This class features load() and save() methods, which will respectively
load and save all the Q_PROPERTY (which have STORED to true) to the
settings INI file.

It also features a js_register() method, which will register the
ApiObject with a JavaScript engine. Then, all the Q_PROPERTY (which have
SCRIPTABLE to true) will be available from JavaScript code.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>