Tool Launcher: use a simple HTML view instead of WebKit
The QtWebKitWidgets module has been deprecated for a long time, and
dropped in Qt 5.7, in favor of QtWebEngine.

However, QtWebEngine packs the full Chromium, and is available only on
x86 and ARM, which severely reduces the platforms that we can support.
Moreover, it causes packaging issues in the upstream Debian project; as
of now, the QtWebEngine is still not available on that distribution (and
probably others).

The thing is, we don't actually need all the features of a web browser
just to display some welcome text and wiki-like internal documentation.
Qt already supports rich text in its widgets as well as HTML 4.
For now, we will just use a simple HTML view to display a welcome
screen, which will be extended later as needed.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>