Osc: Simplify code by moving some things to UI
Instead of creating buttons and check boxes for the cursors,
measurements and trigger settings from the C++ code, we have them inside
the Oscilloscope UI file now.

Also, the push buttons are now using the class CustomPushButton, which
allows the buttons in a button group to be all deselected without any
extra code.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>