PatternGenerator - implemented basic save and restore functionality
Due to the fact that the current save and restore functionality does not
allow the storing of hierarchical structures, a json is generated from
the configuration in the property getter and the json is parsed in the
setter so the pattern generator channels are saved and restored correctly.

Channel group structure, enable, collapse and group status is saved.

Signed-off-by: Adrian Suciu <adrian.suciu@analog.com>