NetworkAnalyzer: Set the graphs' min/max when setting properties
Previouly, the properties were only setting the spin buttons controlling
the min/max values. This caused them to emit a signal, that was caught
by the graphs to update their own setting.

However, in case the value of the property to write was equal to the
current value of a spin button, the signal was not emitted, so the graph
did not update properly its min/max values.

This commit fixes this behaviour.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>