Signal Generator: Don't enforce perfect square waveforms at all times
We first try to find the perfect rate for the waveform (the slowest rate
that we can create a buffer from, where one period corresponds to an
even number of samples, when using oversampling).

If we can't get a perfect rate, we will use the highest rate that we can
create a buffer from, ignoring imperfections ('low/high' samples ratio
not 50/50).

Fixes #80.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>