Calibration: Allow settings written to hardware to take effect
The need for these delays was first observed when the ad5625 channels
had the raw attribute set to values that were very far from the middle
scale (2048). Setting the channels to middle scale and reading data
shortly after that was causing the calibration results to drift from the
ideal results, hence the delays.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>