Osc: Implement automatic switching between low and high gain modes
The high to low switch is done when the current input range of the
device is smaller than the plot height. Without this inputs signals
may be clipped due to ADC overrange.
In other words the input range of the device gets extended in order to
properly display an input signal and avoid ADC saturation.

Since the vertical position and trigger position depend on the gain mode
they will also be updated.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>