Osc: Fix offset handlers not writing offset to hardware
What offset handlers do is that they set the offset of the plot then
they update the value of the vertical position spinbox causing a
callback to execute. The callback was doing nothing because of:
if (value != -plot.VertOffset(current_channel))
since the offset handler had already changed the plot vertical offset.

The offset change was working properly only when it was changed from the
spinbox directly.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>