IIO manager: Emit a signal if the flowgraph errored
There is no way to just query the state of the flowgraph. To work around
this, we start a thread that will just wait() for the flowgraph to end;
and if that event happens, it will send a signal.

When stopping properly, however, the thread is stopped first, the
flowgraph second, so that the signal is not emitted.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>