Osc: Fix time position upper limit not being updated based on time base
When time bases changes it is required the time position to be
verified and if it exceeds the size of the hardware trigger the position
will be adjusted accordingly.
Before the fix, if the time position was large (e.g. seconds with high
sample rate) the application crashed due to buffers size exceeding
Gnuradio buffer size limits.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>