ToolLauncher: The menu is loaded on startup
When the application starts, the menu is loaded. If the user changes the
order of some tools, when closing, the order will be saved and restored
on startup.

Added a dragZone after the last menu option to be able to drag and drop
between the last menu option and the scopy logo.

Signed-off-by: Daniel Guramulta <daniel.guramulta@analog.com>