FftDisplayPlot: Add option to update the sample rate upon new data
Changing the sample rate of the plot while the plot still displays the
data captured with the old sample rate might not be always useful due to
the fact that a call to the FftDisplayPlot::setSampleRate() also
reconfigures the frequency axis making it look like the data that was
originally captured at the old sample rate is now captured at the new
sample rate.

This option gives more flexibility, allowing the sample rate to be
update only when a new set of data is being received by the plot.

Signed-off-by: Dan Nechita <Dan.Nechita@analog.com>