Logic Analyzer: Align buffer size to a multiple of 8 bytes
There is a restriction in the HDL that a buffer size for data capture from
the logic analyzer must be a multiple of 8 bytes. If it is not a multiple
the buffer size is rounded down in the IIO framework to be a multiple of 8.
This is not so much a problem in streaming mode without triggers where the
IIOD will just do multiple captures to get the full data.

In triggered single shot mode on the other hand if the requirement is not
met we end up waiting for the remaining bytes, which will never be
delivered unless a second trigger event happens. This is not the expected
behavior.

Make sure the buffer size is rounded up to the next multiple of 4 samples
(8 bytes).

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>