ToolLauncher: changed the way the detaching works
The detached window will no longer have the tool launcher as a parent,
this will allow to bring the window in front of the main scopy window or
another detached tool.

The run/stop button will now work on the corresponding tool in the
detached window.

Clicking the menu option of the corresponding detached tool will make
the detached window, that holds that tool, come to the front or flash to
know which window it is.

Signed-off-by: Daniel Guramulta <daniel.guramulta@analog.com>