Osc: Fix desynchronization between the channel indexes
Class Oscilloscope updates the index of each channel upon channel
removal/addition so that we always have channel 0, 1, 2, .. even though
the channel names might be Channel0, Channel3, Channel4, etc.

This fix makes sure that class CapturePlot does the same update when it
comes to the set of measurements that it holds.

Signed-off-by: Dan Nechita <Dan.Nechita@analog.com>