SignalGenerator: Fix the glitch when switching from one chn to the other
This is a fix/refactor commit that is based on the recent refactoring
(class Channelwidget).
It fixes the issue that was happening when switching from the menu
button of one channel to the menu button of another channel. The glitch
was that the right menu was being close suddenly instead of being animated.

Signed-off-by: Dan Nechita <Dan.Nechita@analog.com>