Oscilloscope: Fix crash that sometimes occurs when autosetting
The crash happened after setting up a new autoset gnuradio flow, a
singleCaptureDone() event is pending. The event is triggered before the
gnuradio flow actually runs, so no data is available for the autoset
to work with. The fix consists in verifying if there is any data in the
autosetFFTSink vector.

Signed-off-by: Adrian Suciu <adrian.suciu@analog.com>