Minor bug fixes related to the device and how connect/disconnect work
After connecting to a device the user will be able to disconnect only when the calibration is finished, if not skipped.
Fixed some bugs where after removing a device, the left/right menu controls wouldn't work.
Fixed the size of the device in order to fit the green status bar when connected.

Signed-off-by: Daniel Guramulta <Daniel.Guramulta@Analog.com>