CI/travis/before_install_darwin.sh: fix error message during dep install
Now, with `set -e` enabled, brew will fail the build when `brew install`
will error out with "package is already installed" type errors.

To fix this some shell trickery will be used to install first, then if it
fails, it will upgrade, and if that fails, it will check if it is
installed.

The last error code in that construct will be taken into consideration.

Also, the list of deps to be installed has been trimmed here.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>