CI/travis/lib.sh: export PYTHON3 env var
libsigrokdecode seems to try python3.6 to python3.2 commands.
In Travis CI these commands seems to exist as shims, so during configure
`python3.6` will be found, but when using it, it will give an error and
fail the build.

So, just export the python3 command with the PYTHON3 var, and
libsigrokdecode will pick that up.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>