CI/travis/before_install_lib.sh: apply patches only in get functions
Build deps are downloaded, and the patch is applied in the qmake build
function (on each build).
The build (qmake) function is called every time the Scopy build runs, but
the dependencies are downloaded only once, and after that Travis CI caches
them.

It seems that the patch function fails every now and then, since it cannot
re-apply a patch. This wasn't happening before, so, it's probably that some
marker file is not stored properly during caching.

Doing the patching in the getfuncs also removes the necessity to check for
marker files (when patching), since we're more certain that in this case,
the dependencies are patched (if the deps folders already exist).

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>