CI/travis/before_install_linux.sh: use proper paths when building libad9361-iio locally
For DLIBIIO_INCLUDEDIR, the type should be PATH (even though it may not
matter much).
For DLIBIIO_LIBRARIES, the type should be FILEPATH (even though STRING may
be just as fine.
However, for DLIBIIO_LIBRARIES we need to specify the actually libiio.so,
not the library search path.

This only fixes the local build of Scopy on Linux.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>