build,travis-ci: add Xenial builds to the job run
Time to also build for Ubuntu 16.04.
Unfortunately, the QT5 libs are still too old for Scopy, so we need to pull
an external Qt5.9 repo from launchpad (similar as for Trusty).

Link is:
  https://launchpad.net/~beineri/+archive/ubuntu/opt-qt592-xenial

Same packager also has builds for Xenial.

And we also need to specifically pull the Mesa GL dev libs so that libqwt
builds fine.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>