ToolLauncher_API: expose the ability to set/unset AlwaysOnTop mode
Especially when scripting certain mechanisms, a user may want to make sure
that window doesn't go away when accidentally clicking somewhere else (due
to 'reasons').

This change allows for a `launcher.always_on_top = true` behavior to be
specified, and that will make sure the Scopy window stays on top (in a
pretty insistent manner).

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>