LogicAnalyzer: fixes bug that crashes scopy when loading disabled channels
This bug is caused by the fact that the Logic Analyzer and the Logic
Analyzer UI are 2 objects that are synchronized by the call of update_ui()
When loading an .ini file, the LogicAnalyzer channel manager is reconfigured
but the UI is only updated later once the ini file has been loaded. However
the disabled/enabled API attribute changes the UI. If this call happens
before the objects are in sync, the application crashes.

The flag makes sure that the UI update is not done by the API call while
the whole logic analyzer is loading from .ini.


Signed-off-by: Adrian Suciu <adrian.suciu@analog.com>