Voltmeter/SpectrumAnalyzer: don't start the tool if no run/single button is pressed
The tools share an instance of the iio manager for every iio device, in this case the "m2k-adc". iio->started() will return true or false depending if the "m2k-adc" is being used by this tool or another. If the iio device is used by another tool, in some cases, changing some hardware settings might start the flowgraph for a stopped tool. In order to fix this we also check if the run buttons are also checked before doing any operation that might change the state of the flowgraph

Signed-off-by: Daniel Guramulta <daniel.guramulta@analog.com>