Added the phonehome mechanism.
PhoneHome class makes a request to https://analog-applications-versions.herokuapp.com/all
and takes the latest versions of scopy and m2k-firmware.

The server in deployed on heroku and it checks every day for newer scopy and m2k-fw versions.
PhoneHome takes data from here and uses it inside scopy, notifying the user in case a newer version is available.

Signed-off-by: Cristi Iacob <cristian.iacob@analog.com>