Graphical interface for m2k-fw version notifications.
If the phonehome functionality is enabled, a new label will appear at the bottom of each device's menu.
In this label will be displayed whether or not the current m2k-fw version is the latest. If not, there will be a hyperlink displayed, redirecting the user to the wiki page for m2k firmware, describing how to do the update.

Signed-off-by: Cristi Iacob <cristian.iacob@analog.com>