Scopy version checking and message box at first run.
On scopy first run, it will pop this message box asking for automatic version checking.
Starting the checking process.
If the phonehome mechanism is set in the preferences menu, PhoneHome will take the necessary data when launching the application.
Displaying scopy version update.
At the bottom of the home page a label will be displayed. If there is a new scopy version, the label will also provide a link to analog wiki, explaining how to do the update.

Signed-off-by: Cristi Iacob <cristian.iacob@analog.com>