Logic Analyzer: fix blocking when running
When running the LogicAnalyzer with multiple groups enabled, each having a
decoder active, Scopy would freeze and never recover. The problem is that
each instance of DecoderStack (the class that manages decoding) spawns a
thread that would lock a global mutex for the access to libsigrokdecode.
When this decoders are stopped if the thread that currently owns access to
libsigrokdecode is not joined first, the freeze occurs as there is no way
to interrupt that thread from waiting to acquire the mutex, while also not
having a way to interrupt the thread that acquired the mutex. To fix this
we allow each thread to defer the mutex while only one of them gets to own
it at a time. When the threads are now joined, they will stop trying to
acquire the mutex and exit peacefully.

Signed-off-by: Daniel Guramulta <daniel.guramulta@analog.com>