Add a qt.conf file for Windows
This file will specify that Scopy is DPI unaware. This means that Windows
will run Scopy as if it was running on a standard display instead of a High
DPI one. Scopy will still scale accordingly to the Windows display scale
setting but now it will work properly without making the plots slower

Signed-off-by: Daniel Guramulta <daniel.guramulta@analog.com>