General: Update Scopy to make it compatible with Qt 5.15 and the latest Qwt updates.
This manages some missing includes (since Qt 5.15 contains changes in the
"include" system). Also, the qwt_compat is dropped, which only added an
extra layer of definition over Qt types.
This also includes a patch for QwtPolar to make it compatible with the
latest Qwt changes, since this was not done officially.

Signed-off-by: AlexandraTrifan <Alexandra.Trifan@analog.com>