Oscilloscope: fix crash when receiving a constant signal
This bug sporadically occurs when a buffer is constant - has the same
value through the buffer. This was tested by creating a synthetic signal
using a gnuradio::analog::sig_source block to feed in a constant waveform.
When this waveform is fed into the flow, the application crashes because
it computes min and max for the signal and eventually computes a difference
which is 0 in this case. In order to prevent this, I provided some default
values for the histogram computation.

Signed-off-by: Adrian Suciu <adrian.suciu@analog.com>