Windows Installer: Drop the SCOPY_PYTHONPATH environment variable.
We no longer set the env var. Version v1.1.2 will be the last one that
knows how to remove the SCOPY_PYTHONPATH env var from the system at
uninstall. If not uninstalled, then the env var will remain and we need
to provide backwards compatibility. That is why we append it to the end
of the path.

If Scopy was installed using the official installer, it will always use
the hardcoded paths in applicationDirPath(), where the python3.X folder is.
If Scopy is built locally on Windows, the user will set the SCOPY_PYTHONPATH
env var manually and, since that one is appended at the end, the program will
still function properly.

Signed-off-by: Alexandra Trifan <Alexandra.Trifan@analog.com>