Voltmeter: implement selectable / auto gain mode
The gain mode can now be selected for each channel. If the corresponding read value is out of the current range we display an error. If in auto mode, instead of displaying an error we update the gain mode of the channel.The gain mode is displayed under the plot

Signed-off-by: DanielGuramulta <Daniel.Guramulta@analog.com>