General: when handling exceptions don't introduce undefined behaviour
Some exceptions would be handled in a thread. Using this macro that creates a QMessageBox would cause undefined behaviour due to the fact that we are creating GUI elements from another thread. If we can identify that this thread does not correspond to the main thread ie the qapplication's thread we invoke the message box. Due to the fact that "msg.exec()" blocks the main thread we also need to block the current thread, to further prevent more undefined behaviour from being introduced, thus the usage of Qt::BlockingQueuedConnection. If we are on the main thread we can simply create the message box like before

Signed-off-by: DanielGuramulta <Daniel.Guramulta@analog.com>