cmake: qwt and qwtpolar check update
Latest qwt bundles qwtpolar with it. This means that only header files are
distributed for qwtpolar, while the implementation is in the qwt library
This patch assumes that you can either have an older version of qwt with
qwt polar separately - this will link both of the libraries. Or you have
the new version which is bundled. In this case you only need the header file
for qwt polar, and the rest will be bundled in the qwt library

Signed-off-by: Adrian Suciu <adrian.suciu@analog.com>