IioUtil: Add a Command Queue structure to handle IIO commands.
If tinyiiod contexts are used, the IIO calls cannot be done in parallel,
since the daemon can't handle multiple commands at once. This command
queue structure would allow the creation of any Commands and execute them
in the desired manner. The Command Queue can be created to run in the main
thread, run in a work thread and execute commands in place or run in a
work thread and execute each command on a separate thread.

Signed-off-by: AlexandraTrifan <Alexandra.Trifan@analog.com>