Ad74413r: Disable the threshold control while acquisition is in progress.
When the acquisition is in progress, the device is put into a "hardware
mode" which doesn't allow specific attribute reading. If the "threshold"
attribute is read during acquisition, it will become empty which leads
to unexpected behaviour. Also, writing the attribute will not have the
desired output. Thus we disable the threshold control while running for
now. If things change in firmware we will revisit this approach.

Signed-off-by: AlexandraTrifan <Alexandra.Trifan@analog.com>