core: Add a mechanism to print the stacktrace on crash using qDebug.
This mechanism will only provide a stacktrace on platforms that provide
glibc, thus on Windows we will only have the application log without
any stacktrace. For now it should be enough. Research was done on
boost::stacktrace, breakpad also. In the future, Scopy can pe upgraded to
C++ 23 which will provide a mechanism for crash reporting.

Signed-off-by: AlexandraTrifan <Alexandra.Trifan@analog.com>