core/scopymainwindow_api: Changed runScript method.
The method no longer requires the contents of the script as a parameter
and does not automatically close the application.
Closing the application can now be managed by the user both from the
command line through the --keep-running (r) option and from ScopyJS through
runScript(QString scriptPath, bool exitApp) method.

Signed-off-by: andrei.danila <andrei.danila@analog.com>