pqm: add a callback for the PQM ping task and handle the ping internally.
In order to avoid thread concurrency over the serial connection while
acquiring data and also pinging the board, we need to handle the ping
procedure internally. This can only be reproduced on Windows (my guess is
that the different implementation of libserialport for Windows vs UNIX
is the cause for this). Serial port calls return -995 at some point (which
is the error code for no I/O access. After this happens there is no
recovery for the serial connection and the plugin disconnects from the
device. This happens after a random amount of time.

This commit is a temporary workaround due to demand of users. But there is
a task in progress for development and investigation on the IIOPingTask
which should handle the plugin callback for the ping procedure but
should also move the ping management in the DeviceManager rather than
having it inside each plugin.

Signed-off-by: AlexandraTrifan <Alexandra.Trifan@analog.com>