Revert 63 commits from main,the difference between maint-1.4 and maint-1.5 This commit reverts all commits in range 3453f0e and 438f138 (included). These 2 commits mark the interval limits (https://github.com/analogdevicesinc/scopy/tree/maint-1.5 and https://github.com/analogdevicesinc/scopy/tree/maint-1.4).
Scopy2 was developed on a branch that diverged from main at maint-1.4.
During development, fixes for Scopy1 were found and fixed on both "main"
and the Scopy2 "dev" branch thus making it harder now to rebase dev and
main in order for Scopy2 to become the new official main branch.

For maintenance purposes, Scopy1 is kept as maint-1.5, all diverging
commits are reverted on the main branch and the dev branch is rebased
above.


Signed-off-by: AlexandraTrifan <Alexandra.Trifan@analog.com>