core: check if the uri already exists when adding a new device in browser.
If the device was save as selected for auto-connection from a previous
session but the same device is also detected during scanning, the
browser will have duplicate entries of the same device, allowing us
to connect to both. This should not be possible in most cases: the best
example is the ADALM2000 USB connection.

Signed-off-by: AlexandraTrifan <Alexandra.Trifan@analog.com>