m2k: fix the display resample in scenarios with ratio = 0
In different scenarios, the channel combination can generate a zero ratio
when computing the resampling preview parameters. One example is having
a buffer source for both channels and configuring 1kHz and 5MHz. This
configuration can also be tested using the stairstep waveform.

Add a plot status informing the user that the preview is no longer
updated in these scenarios and make sure the gnuradio flow is properly
updated.

Signed-off-by: AlexandraTrifan <Alexandra.Trifan@analog.com>