debugger/iioexplorer: Add code gen feature for debugger plugin
Each selected iio item from the debugger will provide the C code for
accessing and reading it. If multiple items are added to the watchlist,
the code accessing and reading all of them will be generated and can be
saved to a file.

IMPORTANT NOTE: Reading debug device attributes will not work. This is
just a demo.

Signed-off-by: Andrei-Fabian-Pop <Andreifabian.Pop@analog.com>