main: Add QApplication attribute "DontCreateNativeWidgetSiblings"
On Windows, when detaching a widget from an area with 2 widgets,
the plot from said widget will create OpenGL errors. This is
because when detaching a widget it might get promoted to a
native widget (controlled by windows) and thus losing the
"control" of OpenGL. Setting this flag fixes most of these
issues by not allowing the promotion unless absolutely
necessary.

Signed-off-by: Andrei-Fabian-Pop <Andreifabian.Pop@analog.com>